feat(generation): move processing to a dedicated dir
